% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_itens_by_id.R
\name{get_list_itens_by_id}
\alias{get_list_itens_by_id}
\title{Get Tibble from SharePoint List Items by ID}
\usage{
get_list_itens_by_id(site, id)
}
\arguments{
\item{site}{SharePoint site object from \code{get_sharepoint_site("site_name")}}

\item{id}{SharePoint list ID}
}
\value{
A tibble with the list items
}
\description{
Get Tibble from SharePoint List Items by ID
}
\examples{
\dontrun{
site <- get_sharepoint_site("site_name")
get_list_itens_by_id(site, "list_id")
}
}
