% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_properties.R
\name{get_list_properties}
\alias{get_list_properties}
\title{Get Properties of All SharePoint Lists in a Site}
\usage{
get_list_properties(site)
}
\arguments{
\item{site}{SharePoint site object from \code{get_sharepoint_site("site_name")}}
}
\value{
A tidy tibble with list properties
}
\description{
Get Properties of All SharePoint Lists in a Site
}
\examples{
\dontrun{
site <- get_sharepoint_site("site_name")
get_list_properties(site)
}
}
