% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_itens.R
\name{get_site_itens}
\alias{get_site_itens}
\title{Get Items from All SharePoint Lists in a Site}
\usage{
get_site_itens(site, .progress = FALSE)
}
\arguments{
\item{site}{SharePoint site object from \code{get_sharepoint_site("site_name")}}

\item{.progress}{Logical, whether to show progress bar}
}
\value{
A tibble with one row per list and a nested column with items
}
\description{
Get Items from All SharePoint Lists in a Site
}
\examples{
\dontrun{
site <- get_sharepoint_site("site_name")
get_site_itens(site)
}
}
