% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_prop_to_tibble.R
\name{list_prop_to_tibble}
\alias{list_prop_to_tibble}
\title{Convert SharePoint List Properties to Tibble}
\usage{
list_prop_to_tibble(list)
}
\arguments{
\item{list}{A list of SharePoint list properties, typically from \code{get_sharepoint_site("site_name")$get_lists()[[1]]$properties}}
}
\value{
A tibble with nested list column
}
\description{
Convert SharePoint List Properties to Tibble
}
