% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_list_properties.R
\name{tidy_list_properties}
\alias{tidy_list_properties}
\title{Tidy SharePoint List Properties Tibble}
\usage{
tidy_list_properties(tbl)
}
\arguments{
\item{tbl}{Tibble from \code{list_prop_to_tibble()}}
}
\value{
A tidy tibble with flattened and parsed datetime columns
}
\description{
Tidy SharePoint List Properties Tibble
}
